requirements-dev.txt
black==25.1.0
flake8==7.3.0
isort==6.0.1
mypy==1.17.1
django-stubs==5.1.0
django-stubs-ext==5.1.0
